feat: Create and use classes
In this exercise you'll be creating your own custom task class. Instead of
configuring ad-hoc tasks in a build script, we can define a task as a regular
Java or Groovy class that can then be reused throughout your build scripts.
Remember, a Gradle build script is just a Groovy script, so you can write your
class definition in-line in the script itself.

Exercises:

1. Define a new task class named "HelloTask"
2. Define a task action for your new task that simply prints the string
   "Executing my custom task." to the console.
3. Create a new task in your build script of type HelloTask named "hello".
4. Run the "hello" task and observe the output.