feat: Declare Repository & Dependencies
In this set of exercises we'll be working with Gradle's dependency management
features. Specifically, we'll be declaring repositories and dependencies for
our project. We've already applied the Java plugin, so no need to do that.
Also, we've supplied a simple task you can use to see all the dependencies
that have been configured for the 'compile' configuration.

Exercises:

1. Configure Maven Central as a repository
2. Add a dependency to the 'compile' configuration for the latest version
   of commons-io.
3. Add a file dependency that includes all the jars in the `libs` directory.
4. Run the `printDependencies` task to verify the dependencies.
5. Run the `printDependencies` task with the --refresh-dependencies flag.
6. Run the `printDependencies` task one more time and note the difference in
   run time.

Hint: You can search for available dependencies in Maven Central online.
Use the following website to find the correct group, name and version
for commons-io.

http://search.maven.org/