fix: Configure the Gradle Wrapper
Had trouble getting to use gradle wrapper with AS

In this series of exercises we'll be practicing initializing and configuring
the Gradle wrapper. In this case you are given an empty project. You will use
your installed version of Gradle to "bootstrap" this project by adding the
wrapper files.

Exercises:

1. Run the 'wrapper' task to create the Gradle wrapper files.
2. Use the wrapper to run the 'tasks' task.
3. Configure the 'wrapper' task to use Gradle version 2.14.1.
4. Use the wrapper with the '--version' option to confirm it is using
   Gradle version 2.14.1.
5. Modify the gradle-wrapper.properties file to use the latest Gradle
   nightly build.

Hint: For more information on using Gradle nightly builds, check out the
Gradle website: http://gradle.org/gradle-nightly-build/