feat: Declare Flavors
Install on device
In this series of exercises we'll be experimenting with Android product
flavors. In this exercise you'll be implementing one of the most common
product flavor use cases, free and premium app variants. You'll be asked to
create two new product flavors. Some additional resources are already provided
for each product flavor.

Exercises:

1. Import the project into Android Studio.
2. Build the project and deploy it onto a device (or emulator). Note the
   product flavor shown on the main activity.
3. Define two new product flavors, 'free' and 'paid'.
4. Add new resource files so that the paid flavor tells a funny joke, and the
   free version tells a cheesy joke.
5. Use the 'Build Variants' window in Android Studio to select each new variant
   and build/deploy the application onto a device.
6. Confirm that the correct product flavor is shown on the main activity.