fix: This is not working.  Look at solution for correct answer.
You were close, but was about to get a little off the rails.  Should
try again later.  But...

You learned that if you wat to have code that will be a little different in 2
different flavors, you have to remove that code from the 'main' source set
and put that code in the other 2. Change it how you need to.

If you have a flavor that has it's own code that only needs to be run there
you will have to make a manifest specific to that.  Do not change the one in
'main'.  Also you need to put it in a folder with the same name as the flavor
at the end of the uri.  Like, paid/java/com/udacity/gradle/flavorspecificactivity
/paid .

Unlike code, strings.xmls and Manifests are merged.