feat: Create a Java Library and use it
The settings.gradle file is where we tell Gradle what subprojects to include
in the multi-project build. We already have the :app project for the Android
app, and we'll add :javaJokes

Now javaJokes is a subproject, and if we run

    $ ./gradlew projects

We see that both app and javaJokes are subprojects.

Finally, if we want to use code from javaJokes in our app project, we need to
declare a project dependency. Let's head over the build.gradle file in the app
directory.

---
Now we need to declare a compile time dependency on the project :javaJokes

This bit is a little tricky. The Java plugin defines an extra "default"
configuration we haven't talked about yet. It adds its output: its JAR, to
that configuration. When we declare our dependency on the :javaJokes project,
we're actually depending on the "default" configuration of :javaJokes, which
contains the output JAR.

---
Finally, we need to actually use the library. Give the textView in
fragment_main.xml an id, we used textView. Then in MainActivityFragment, add:

    import com.udacity.gradle.jokes.Joker;

Then replace onCreateView in MainActivityFragment with the following.

    @Override
    public View onCreateView(LayoutInflater inflater, ViewGroup container,
                             Bundle savedInstanceState) {
        View rootView = inflater.inflate(R.layout.fragment_main, container, false);
        TextView textView = (TextView) rootView.findViewById(R.id.textView);
        Joker myJoker = new Joker();
        textView.setText(myJoker.getJoke());
        return rootView;
    }

Then we're ready to build! When you run the build, watch the Gradle Console,
which you can open at the bottom right of Android Studio. You'll notice
something like this:

:app:preBuild UP-TO-DATE
:app:preDebugBuild UP-TO-DATE
:app:compileDebugNdk UP-TO-DATE
:app:checkDebugManifest
:app:preReleaseBuild UP-TO-DATE
:javaJokes:compileJava
:javaJokes:processResources UP-TO-DATE
:javaJokes:classes
:javaJokes:jar
:app:prepareComAndroidSupportAppcompatV72210Library UP-TO-DATE
:app:prepareComAndroidSupportSupportV42210Library UP-TO-DATE
:app:prepareDebugDependencies

The app project was running some tasks, then the javaJokes project jumped in
to create a JAR for the app project to use!

That's all there is to it to add a Java library project to an Android Studio build!