feat: Create 2 Java Libraries
Time for you to modularize your joke telling app. You'll create a joke telling
app, and not one, but two Java libraries!

This one is very important

1. Create a new project in Android Studio, using all the defaults.

2. Manually add a Java module with a class called JokeSmith with a getJoke()
method. Feel free to supply a funny joke, or just a debug method, if
you're no fun. Remember to set sourceCompatibility = 1.7, so your library is
usable on Android. This line should be added to the build.gradle file of the
module.

3. Add your new module to your settings.gradle file

4. Declare a project dependency between your app and JokeSmith

5. Add a text view to MainActivityFragment and populate it with a joke from
JokeSmith.

6. Use the Android Studio wizard to add a second Java library called
JokeWizard. To access the wizard, navigate to:

    File > New > New Module > Java Library

Note that the wizard adds JokeWizard to your settings.gradle

7. Declare a project dependency between JokeWizard and JokeSmith

8. Implement a getJoke() method on JokeWizard that uses the getJoke() method
on JokeSmith.

9. Add a project dependence between your app and JokeWizard

9. Add another text view to MainActivityFragment and populate it with a joke
from JokeWizard.