feat: Part of making Android Library lesson; make Java Libary that tells joke
Let's further break up the functionality of the Joke telling app by creating a
new activity to display the joke, and moving that activity to an Android
Library project.

1. Create a new project in Android Studio, using all the defaults.

2. Create or reuse a Java library that can supply a joke.

- You Created the folder structure and the files for the .java file and
build.gradle in a terminal using nano. Then you put ':chuckles' in the
settings.gradle and called, [gradle build]. You also used mkdir to get
from the chuckles folder to the java file. But you are not sure how to
know what path should be. So just used the normal package name.