feat: Start to set up Unit Testing
Create the folders and Classes necessary to to set up testing.
Currently, you got all the red squigglies to disappear and
everything seems to know that it exists, but you have introduced
duplicates somewhere in the build.gradle. There is an extra thing
in your code due to transitive dependency. You have to remove the
extra and different dependency so that the conflict can be resolved
you can run the android tests. You will have to do this in a future
commit.