feat: Finish setting up unit testing
You have to lower the version number of the 'rules' and
the 'runner' from 1.02, to 0.5. So that its associated
appcompat matches the version of 'espresso-core' you are using.

In this exercise you will set up unit tests and connected tests for a simple
Android app.

Exercises:

    1. Set up unit tests that verify the initial count and increment methods
    of the ClickCounter class
    2. Set up connected tests that test that clicking on the button actually
    increments the value displayed

To complete the latter task, check out:

    https://developer.android.com/training/testing/ui-testing/espresso-testing.html

And watch out for misspellings when things do not work out.

Look this up later: '
androidTestImplementation('com.android.support.test.espresso:espresso-core:2.2.2', {
        exclude group: 'com.android.support', module: 'support-annotations'
    })'