feat: Create a Task Rule
This is very interesting. Create a Task Rule that will be able
to turn any folder in the 'src' directory into an archive under
a new, 'build' directory.

In this exercise you'll define a task rule for your Gradle build script. We're
using the same source files from the previous few exercises again. Task rules
are a create way to define dynamic tasks. In this example, you'll be defining
a task rule which will bundle our various source collections.

Exercises:

1. Define a new task rule that matches a task name pattern of "bundle<Source>".
2. Task rule should define a new Zip task with the following properties:
    a) Archive name should be "<Source>.zip"
    b) Destination of archive should be "build" directory.
    c) Archive includes all files from "src/<Source>" directory.

*/