feat: Create an Android Library
Create an Android Library containing an Activity that will display a joke
passed to it as an intent extra. Wire up project dependencies so that the
button can now pass the joke from the Java Library to the Android Library.

For review on how to create an Android library, check out demo 4.03. For a
refresher on intent extras, check out;

http://developer.android.com/guide/components/intents-filters.html

You want to add a test. To see if it was working.
This will help to figure out how to test an Activity waiting for an intent.
Since we are an internal app, and are just called with an Intent fron another
app.
https://xebia.com/blog/android-intent-extras-espresso-rules/
https://github.com/barend/android-espresso-intent-sample

Weird: Command Line connectedCheck asked for monitor:1.0.2 but did not find it.
It seems to add Theme.
       IDE Gradle connectedCheck asks for Theme in order to work. Added a theme to Manifest.
       Sometimes when both solutions are in, will say there are 2 themes but sometimes works.
       If having issues, remove one or the other and put back in one at a time