fix: Add code to test that text is displayed on joke showing Activity
Add code to test that your Async task successfully retrieves a non-empty
string. For a refresher on setting up Android tests, check out demo 4.09.

Create an androidTest directory in correct location. Make sure you are
in the right one, or will not have access to the appropriate functions.
Make sure to have all the correct Espresso stuff in the build.gradle for
the part you are testing.