feat: Add Free and Paid flavors to app
You will have to remove a lot of dependencies (real, not Gradle specific)
and realize that Values and Manifests are merged, while Java classes cannot
be merged. So, The code closest to the core will get to the edges, and code
for specific things will need to be put higer up. Like the another Android
Manifest for Free, but not for Paid. Because, the one in main will be
sufficient. Also the Fragment java code will need to be removed from main
and placed where necessary with appropriate changes if necessary.
Aside: Realize that if you want to have a Values file in a variant you
have to create the file; not, the folder. That will be frustrating.