feat: Start getting ready for ultimate testing tasks
Start by getting any of the tasks asked for, started in the first place.
So, can now start appEngine, and stop it, and run connected checks after this.
But, have yet to fix the testing problem created by the Interstitial ad.
Will have to fix that later by testing Intents instead of actually moving
from one Activity to another.