feat: Get tests working with Ads and Interstitial
You may find that you cannot press the button and get a reaction if the ad at
the bottom of the screen has not appeared yet. Try using a Thread.sleep() or
Idling resource, later. It is also important to know that to exit an ad, you
can press the, 'X', or press back. Pressing back will not affect the flow of
the app.