fix: Change XML file to get all of words to fit on background
You should change what you are using to contain the text. You should also
change the width to 0dp since you do not know how long they'll be ahead of
time. And 0dp in a ConstraintLayout is matches_constraint. MP is not really
supposed to be used. So use startToStartOf/EndToEndOf and 0dp to contain text.